+++ examples/app-store-apps/1-basic-launchpad-app/astro.config.mjs
@@ -3,5 +3,6 @@ import { defineConfig } from 'astro/config';
 import netlify from '@astrojs/netlify';
 
 export default defineConfig({
+  output: 'server',
   adapter: netlify()
 });