+++ packages/plugins/didkit-plugin-node/check-exports.cjs
@@ -1,0 +1,3 @@
+const native = require('@learncard/didkit-plugin-node');
+console.log('Native module exports:', Object.keys(native));
+console.log('Full native object:', native);