+++ packages/react-learn-card/src/components/MeritBadgeDisplayCard/MeritBadgeFrontFace.tsx
@@ -39,6 +39,9 @@ type MeritBadgeFrontFaceProps = {
     showDetailsBtn?: boolean;
     formattedDisplayType?: string;
     customBodyContentSlot?: React.ReactNode;
+    unknownVerifierTitle?: string;
+    hideAwardedTo?: boolean;
+    hideFrontFaceDetails?: boolean;
 };
 
 export const MeritBadgeFrontFace: React.FC<MeritBadgeFrontFaceProps> = ({
@@ -56,6 +59,9 @@ export const MeritBadgeFrontFace: React.FC<MeritBadgeFrontFaceProps> = ({
     showDetailsBtn = false,
     formattedDisplayType,
     customBodyContentSlot,
+    unknownVerifierTitle,
+    hideAwardedTo: hideAwardedToProp,
+    hideFrontFaceDetails,
 }) => {
     const {
         title = '',
@@ -108,10 +114,14 @@ export const MeritBadgeFrontFace: React.FC<MeritBadgeFrontFaceProps> = ({
     const isAppIssuerDid = issuerDid?.includes(':app:');
 
     let verifierState: VerifierState;
+    const hideAwardedTo =
+        hideAwardedToProp ?? (issueeName?.includes('did:key') || issueeName?.includes('did:example:123'));
     if (credentialSubject?.id === issuerDid && issuerDid && issuerDid !== 'did:example:123') {
         // the extra "&& issuerDid" is so that the credential preview doesn't say "Self Verified"
         // the did:example:123 condition is so that we don't show this status from the Manage Boosts tab
         verifierState = VERIFIER_STATES.selfVerified;
+    } else if (unknownVerifierTitle) {
+        verifierState = VERIFIER_STATES.trustedVerifier;
     } else {
         if (knownDIDRegistry?.source === 'trusted') {
             verifierState = VERIFIER_STATES.trustedVerifier;
@@ -241,6 +251,7 @@ export const MeritBadgeFrontFace: React.FC<MeritBadgeFrontFaceProps> = ({
             <div className="absolute bottom-0 w-[calc(100%-26px)] flex justify-center">
                 <VerifierStateBadgeAndText
                     verifierState={verifierState}
+                    unknownVerifierTitle={unknownVerifierTitle}
                     className="bg-white px-[5px] pb-[5px]"
                 />
             </div>